display the video component on the page.
the video is set through dangerouslySetInnerHTML because the error 'Invariant Violation: ReactMount: Two valid but unequal nodes with the same `data-reactid`' keeps showing up when using React to render the video.  My guess is that the library, video.js, is doing some DOM manipulation on the video on mouse over.  Since the DOM is being manipulated, React doesn't like it.  Setting the inner HTML will take the element outside of React, which leaves it safe for a 3rd party to manipulate.

The video component is ugly right now, too.  Styling will come later.